Moved main blueprint into its own folder.
